requirements.txt
streamlit
groq
supabase
pandas
plotly
gtts
sqlalchemy
psycopg2-binary
Pillow
selenium
webdriver-manager
pytest
duckduckgo-search
razorpay
tavily-python
ruff